requirements.txt
Flask>=2.3,<3.1           # Обновляем Flask и задаем верхнюю границу
Flask-Login>=0.6,<0.7
Flask-SocketIO>=5.3,<5.4
Flask-WTF>=1.1,<1.3      # Добавлено для CSRF и форм
python-dotenv>=1.0,<2.0 # Добавлено для .env
pandas>=2.0,<2.3         # Обновляем pandas
requests>=2.31,<2.33
lxml>=4.9,<5.3
plotly>=5.18,<5.23
cssselect>=1.2,<1.3
Werkzeug>=2.3,<3.1       # Зависимость Flask
Jinja2>=3.1,<3.2         # Зависимость Flask
itsdangerous>=2.1,<2.2   # Зависимость Flask
click>=8.1,<8.2          # Зависимость Flask
gunicorn>=21.2,<22.1     # (Опционально, для production)
python-engineio>=4.8,<4.10 # Зависимость Flask-SocketIO
python-socketio>=5.10,<5.12 # Зависимость Flask-SocketIO
gevent>=23.9,<24.1      # (Опционально, для SocketIO async_mode='gevent')
gevent-websocket>=0.10,<0.11 # (Опционально, для SocketIO async_mode='gevent')+
python-dateutil